Force tenant selection if tenant context is invalid
Redirects to tenant selection if the user navigates to a tenant for
which they do not belong or if the tenant does not exist. (+1 squashed commit)
Squashed commits:
[13afb33] Force tenant selection if tenant context is invalid

Redirects to tenant selection if the user navigates to a tenant for
which they do not belong or if the tenant does not exist.